requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
black==23.11.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.3
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
fastapi==0.104.1
filelock==3.13.1
FLAML==2.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.12.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.1
google-search-results==2.4.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==28.1.0
langchain==0.0.346
langchain-core==0.0.10
langsmith==0.0.69
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.2.4
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.1.0
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
